# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProManageX Enterprise SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /violations:
    get:
      operationId: listViolations
      tags: [violations]
      summary: List violations
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of violations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"
    post:
      operationId: createViolation
      tags: [violations]
      summary: Register a violation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateViolationBody"
      responses:
        "201":
          description: Created violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
  /violations/{id}:
    get:
      operationId: getViolation
      tags: [violations]
      summary: Get violation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Violation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
    patch:
      operationId: updateViolation
      tags: [violations]
      summary: Update violation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateViolationBody"
      responses:
        "200":
          description: Updated violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
...
  /work-areas:
    get:
      operationId: listWorkAreas
      tags: [workAreas]
      summary: List work areas
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of work areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkArea"
    post:
      operationId: createWorkArea
      tags: [workAreas]
      summary: Create a work area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkAreaBody"
      responses:
        "201":
          description: Created work area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkArea"
  /work-areas/{id}:
    get:
      operationId: getWorkArea
      tags: [workAreas]
      summary: Get work area by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work area details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkArea"
    patch:
      operationId: updateWorkArea
      tags: [workAreas]
      summary: Update work area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkAreaBody"
      responses:
        "200":
          description: Updated work area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkArea"
    delete:
      operationId: deleteWorkArea
      tags: [workAreas]
      summary: Delete work area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EQUIPMENT ────────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Violation:
      type: object
      required: [id, userId, category, severity, status, date, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: [string, 'null']
        category:
          type: string
          enum: [safety, attendance, conduct, quality, policy]
        severity:
          type: string
          enum: [minor, moderate, major, critical]
        status:
          type: string
          enum: [open, under_review, resolved, escalated]
        date:
          type: string
          format: date
        description:
          type: string
        penalty:
          type: ["number", "null"]
        evidenceUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateViolationBody:
      type: object
      required: [userId, category, severity, date, description]
      properties:
        userId:
          type: integer
        category:
          type: string
        severity:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        penalty:
          type: number
    UpdateViolationBody:
      type: object
      properties:
        status:
          type: string
        penalty:
          type: number
        description:
          type: string
...
    WorkArea:
      type: object
      required: [id, name, location, status, progress, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        description:
          type: [string, 'null']
        supervisorId:
          type: ["integer", "null"]
        supervisorName:
          type: [string, 'null']
        status:
          type: string
          enum: [active, inactive, completed, on_hold]
        progress:
          type: number
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        teamSize:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    CreateWorkAreaBody:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        location:
          type: string
        description:
          type: string
        supervisorId:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
    UpdateWorkAreaBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        description:
          type: string
        supervisorId:
          type: integer
        status:
          type: string
        progress:
          type: number

    # ── Equipment Schemas ───────────────────────────────────────────────────────
...